feat(moral-topology): round 3 — adaptive thresholds, substrate study, consciousness coupling
Adaptive self-tuning anomaly thresholds (EMA-based drift/FE tracking),
consciousness engine module split (mod/types/measure/helpers), substrate ×
moral topology interaction study, mesh gossip wiring (PayloadType::MoralTopology),
anomaly response actions, paper supplement Section S2, soak test + E2E test.

32/32 moral topology tests, 76 consciousness engine tests, 3866/3867 lib tests.

+++ symthaea/src/cognitive_loop/accessors/consciousness.rs
@@ -56,7 +56,7 @@ impl CognitiveLoopService {
             flow_learning_boost: self.flow_state.learning_boost,
             boredom: self.curiosity_drive.boredom,
             curiosity: self.curiosity_drive.curiosity,
-            exploration_urge: self.curiosity_drive.exploration_urge,
+            exploration_urge: self.curiosity_drive.exploration_urge as f32,
             exploring: self.curiosity_drive.should_explore(),
             novelty_bonus: self.curiosity_drive.novelty_bonus,
             emotional_valence: self.emotion_contagion.smoothed_valence(),
